requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
colorama==0.4.4
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==4.0.2
django-binary-database-files==1.0.15
django-heroku==0.3.1
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.0.12
gunicorn==20.1.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psycopg2==2.9.3
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
scipy==1.6.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
tzdata==2021.5
virtualenv==20.4.7
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
xlwt==1.3.0
